requirement.txt
Django==3.1.3
djangorestframework==3.12.2
django-rest-swagger==2.2.0
django-rest-framework-mongoengine==3.4.1
django-cors-headers==3.5.0
pymongo==3.11.1
mongoengine==0.21.0
bcrypt==3.2.0
blinker==1.4
PyJWT==1.7.1